+++ internal/notify/manager.go
@@ -77,6 +77,7 @@ func (m *Manager) parseTemplates() {
 		events.Event_CHANNEL_CLOSE:   m.cfg.Templates.ChannelClose,
 		events.Event_INVOICE_SETTLED: m.cfg.Templates.InvoiceSettled,
 		events.Event_FAILED_HTLC:     m.cfg.Templates.FailedHtlc,
+		events.Event_KEYSEND:         m.cfg.Templates.Keysend,
 	}
 
 	for name, text := range templates {